backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain 相关
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# 向量存储
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# 文档处理
pypdf==3.17.1
python-docx==1.1.0

# 数据处理
numpy==1.24.3
pandas==2.0.3

# 环境变量
python-dotenv==1.0.0

# 工具库
tqdm==4.66.1
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# OpenAI
openai==1.6.1

# 数据库
SQLAlchemy>=2.0.30
PyMySQL>=1.1.0


